requirements.txt
click==6.7
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.3
django-bootstrap-form==3.2.1
django-bootstrap3==9.0.0
django-cors-headers==2.1.0
django-datetime-widget==0.9.3
django-forms-bootstrap==3.1.0
django-zurb-foundation-rtl==6.0.6.3
djangorestframework==3.6.3
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
jpeg4py==0.1.2
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.2
Pygments==2.2.0
python-decouple==3.0
pytz==2017.2
reportlab==3.4.0
static3==0.7.0
virtualenv==15.1.0
virtualenvwrapper-win==1.2.3
Werkzeug==0.12.2
whitenoise==3.3.1
